Add CloudKit-inspired changes fetcher
CDTFetchChanges can be thought a more useful changes feed for
the outside world. It returns either the winning revision for
changes or new documents, or the IDs for deleted documents,
sparing callers from the plumbing required to work this out.

The CDTFetchChanges class allows users to find out what has
changed in a database. Inspired by CloudKit's fetch operations,
it uses a block-based method for notifying of the changes to
a database. The class uses sequence values, as one would expect,
to do this.

The class was build in response to our own issues reading
changes for our indexing code.

BugId: 45845