TD_Database:getDocsWithIDs:Options: Apply LIMIT & OFFSET only to table docs
Previous SELECT joined docs and revs and applied a LIMIT and an OFFSET, if
the LIMIT was one, we only got the first revision for the first document. However,
we really wanted to get the winning revision for the first document, therefore we
changed this method to get all the revision for a document after LIMIT and OFFSET
is applied only to docs.